+++ scout/server/blueprints/cases/templates/cases/cases.html
@@ -12,14 +12,6 @@
   <li class="nav-item active">
     <a class="nav-link" href="#">{{ institute.display_name }} Cases</a>
   </li>
-  {% if config.SHOW_CAUSATIVES %}
-    <li class="nav-item">
-      <a class="nav-link" href="{{ url_for('cases.causatives', institute_id=institute._id) }}">Causatives</a>
-    </li>
-  {% endif %}
-  <li class="nav-item">
-    <a class="nav-link" href="{{ url_for('cases.gene_variants', institute_id=institute._id, variant_type=['clinical'], rank_score=15) }}">All SNVs and INDELs</a>
-  </li>
   <li class="nav-item">
     <a class="nav-link" href="{{ url_for('overview.clinvar_submissions', institute_id=institute._id) }}">Clinvar submissions</a>
   </li>
