+++ scout/server/blueprints/cases/templates/cases/phenotype.html
@@ -32,7 +32,7 @@
             </select>
           </div>
             <button class="btn btn-outline-secondary form-control" type="submit">Add</button>
-            <a href="{{ url_for('overview.institute', institute_id=institute._id) }}">Add options</a>
+            <a href="{{ url_for('overview.institute_settings', institute_id=institute._id) }}">Add options</a>
         </div>
       </form>
     </div>
@@ -151,7 +151,7 @@
           </div>
           <div>
             <button class="btn btn-outline-secondary form-control" type="submit">Add</button>
-            <a href="{{ url_for('overview.institute', institute_id=institute._id) }}">Add options</a>
+            <a href="{{ url_for('overview.institute_settings', institute_id=institute._id) }}">Add options</a>
           </div>
         </div>
       </form>
