+++ scout/server/blueprints/institutes/controllers.py
@@ -4,9 +4,11 @@ import logging
 
 LOG = logging.getLogger(__name__)
 
+from flask_login import current_user
 from flask import flash
 from scout.parse.clinvar import clinvar_submission_header, clinvar_submission_lines
 from scout.server.extensions import store
+from scout.server.utils import user_institutes
 
 
 def institute(store, institute_id):
@@ -46,9 +48,7 @@ def populate_institute_form(form, institute_obj):
     form.frequency_cutoff.default = institute_obj.get("frequency_cutoff")
 
     # collect all available default HPO terms
-    default_phenotypes = [
-        choice[0].split(" ")[0] for choice in form.pheno_groups.choices
-    ]
+    default_phenotypes = [choice[0].split(" ")[0] for choice in form.pheno_groups.choices]
     if institute_obj.get("phenotype_groups"):
         for key, value in institute_obj["phenotype_groups"].items():
             if not key in default_phenotypes:
@@ -86,9 +86,7 @@ def update_institute_settings(store, institute_obj, form):
 
     for pheno_group in form.getlist("pheno_groups"):
         phenotype_groups.append(pheno_group.split(" ,")[0])
-        group_abbreviations.append(
-            pheno_group[pheno_group.find("( ") + 2 : pheno_group.find(" )")]
-        )
+        group_abbreviations.append(pheno_group[pheno_group.find("( ") + 2 : pheno_group.find(" )")])
 
     if form.get("hpo_term") and form.get("pheno_abbrev"):
         phenotype_groups.append(form["hpo_term"].split(" |")[0])
@@ -112,6 +110,99 @@ def update_institute_settings(store, institute_obj, form):
     return updated_institute
 
 
+def gene_variants(store, variants_query, institute_id, page=1, per_page=50):
+    """Pre-process list of variants."""
+    # We need to call variants_collection.count_documents here
+    variant_count = variants_query.count()
+    skip_count = per_page * max(page - 1, 0)
+    more_variants = True if variant_count > (skip_count + per_page) else False
+    variant_res = variants_query.skip(skip_count).limit(per_page)
+
+    my_institutes = set(inst["_id"] for inst in user_institutes(store, current_user))
+
+    variants = []
+    for variant_obj in variant_res:
+        # Populate variant case_display_name
+        variant_case_obj = store.case(case_id=variant_obj["case_id"])
+        if not variant_case_obj:
+            # A variant with missing case was encountered
+            continue
+        case_display_name = variant_case_obj.get("display_name")
+        variant_obj["case_display_name"] = case_display_name
+
+        # hide other institutes for now
+        other_institutes = set([variant_case_obj.get("owner")])
+        other_institutes.update(set(variant_case_obj.get("collaborators", [])))
+        if my_institutes.isdisjoint(other_institutes):
+            # If the user does not have access to the information we skip it
+            continue
+
+        genome_build = variant_case_obj.get("genome_build", "37")
+        if genome_build not in ["37", "38"]:
+            genome_build = "37"
+
+        # Update the HGNC symbols if they are not set
+        variant_genes = variant_obj.get("genes")
+        if variant_genes is not None:
+            for gene_obj in variant_genes:
+                # If there is no hgnc id there is nothin we can do
+                if not gene_obj["hgnc_id"]:
+                    continue
+                # Else we collect the gene object and check the id
+                if gene_obj.get("hgnc_symbol") is None or gene_obj.get("description") is None:
+                    hgnc_gene = store.hgnc_gene(gene_obj["hgnc_id"], build=genome_build)
+                    if not hgnc_gene:
+                        continue
+                    gene_obj["hgnc_symbol"] = hgnc_gene["hgnc_symbol"]
+                    gene_obj["description"] = hgnc_gene["description"]
+
+        # Populate variant HGVS and predictions
+        gene_ids = []
+        gene_symbols = []
+        hgvs_c = []
+        hgvs_p = []
+        variant_genes = variant_obj.get("genes")
+
+        if variant_genes is not None:
+            functional_annotation = ""
+
+            for gene_obj in variant_genes:
+                hgnc_id = gene_obj["hgnc_id"]
+                gene_symbol = gene(store, hgnc_id)["symbol"]
+                gene_ids.append(hgnc_id)
+                gene_symbols.append(gene_symbol)
+
+                hgvs_nucleotide = "-"
+                hgvs_protein = ""
+                # gather HGVS info from gene transcripts
+                transcripts_list = gene_obj.get("transcripts")
+                for transcript_obj in transcripts_list:
+                    if (
+                        transcript_obj.get("is_canonical")
+                        and transcript_obj.get("is_canonical") is True
+                    ):
+                        hgvs_nucleotide = str(transcript_obj.get("coding_sequence_name"))
+                        hgvs_protein = str(transcript_obj.get("protein_sequence_name"))
+                hgvs_c.append(hgvs_nucleotide)
+                hgvs_p.append(hgvs_protein)
+
+            if len(gene_symbols) == 1:
+                if hgvs_p[0] != "None":
+                    hgvs = hgvs_p[0]
+                elif hgvs_c[0] != "None":
+                    hgvs = hgvs_c[0]
+                else:
+                    hgvs = "-"
+                variant_obj["hgvs"] = hgvs
+
+            # populate variant predictions for display
+            variant_obj.update(predictions(variant_genes))
+
+        variants.append(variant_obj)
+
+    return {"variants": variants, "more_variants": more_variants}
+
+
 def clinvar_submissions(store, institute_id):
     """Get all Clinvar submissions for a user and an institute"""
     submissions = list(store.clinvar_submissions(institute_id))
@@ -130,17 +221,13 @@ def update_clinvar_submission_status(store, request, institute_id, submission_id
     update_status = request.form.get("update_submission")
 
     if update_status in ["open", "closed"]:  # open or close a submission
-        store.update_clinvar_submission_status(
-            institute_id, submission_id, update_status
-        )
+        store.update_clinvar_submission_status(institute_id, submission_id, update_status)
     if update_status == "register_id":  # register an official clinvar submission ID
         result = store.update_clinvar_id(
             clinvar_id=request.form.get("clinvar_id"), submission_id=submission_id,
         )
     if update_status == "delete":  # delete a submission
-        deleted_objects, deleted_submissions = store.delete_submission(
-            submission_id=submission_id
-        )
+        deleted_objects, deleted_submissions = store.delete_submission(submission_id=submission_id)
         flash(
             f"Removed {deleted_objects} objects and {deleted_submissions} submission from database",
             "info",
@@ -157,12 +244,9 @@ def update_clinvar_sample_names(store, submission_id, case_id, old_name, new_nam
         old_name(str): old name of an individual in case data
         new_name(str): new name of an individual in case data
     """
-    n_renamed = store.rename_casedata_samples(
-        submission_id, case_id, old_name, new_name
-    )
+    n_renamed = store.rename_casedata_samples(submission_id, case_id, old_name, new_name)
     flash(
-        f"Renamed {n_renamed} case data individuals from '{old_name}' to '{new_name}'",
-        "info",
+        f"Renamed {n_renamed} case data individuals from '{old_name}' to '{new_name}'", "info",
     )
