+++ scout/server/blueprints/institutes/forms.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from flask_wtf import FlaskForm
+from wtforms.widgets import TextInput
 from wtforms import (
     IntegerField,
     SelectMultipleField,
@@ -7,6 +8,7 @@ from wtforms import (
     DecimalField,
     TextField,
     validators,
+    Field
 )
 from scout.constants import PHENOTYPE_GROUPS
 
@@ -55,3 +57,31 @@ class InstituteForm(FlaskForm):
     )
     institutes = NonValidatingSelectMultipleField("Institutes to share cases with", choices=[])
     submit_btn = SubmitField("Save settings")
+
+# make a base class or other utility with this instead..
+class TagListField(Field):
+    widget = TextInput()
+
+    def _value(self):
+        if self.data:
+            return ", ".join(self.data)
+
+        return ""
+
+    def process_formdata(self, valuelist):
+        if valuelist:
+            self.data = [x.strip() for x in valuelist[0].split(",") if x.strip()]
+        else:
+            self.data = []
+
+class GeneVariantFiltersForm(FlaskForm):
+    """Base FiltersForm for SNVs"""
+
+    variant_type = SelectMultipleField(choices=[("clinical", "clinical"), ("research", "research")])
+    hgnc_symbols = TagListField("HGNC Symbols/Ids (case sensitive)")
+    filter_variants = SubmitField(label="Filter variants")
+    rank_score = IntegerField()
+    phenotype_terms = TagListField("HPO terms")
+    phenotype_groups = TagListField("Phenotype groups")
+    similar_case = TagListField("Phenotypically similar case")
+    cohorts = TagListField("Cohorts")
