+++ scout/server/blueprints/institutes/templates/overview/utils.html
@@ -1,42 +1,3 @@
-{% macro institute_actionbar(institute) %}
-<!-- Collapsible Sidebar, Based on https://www.codeply.com/go/LFd2SEMECH -->
-  <div id="sidebar-container" class="sidebar-expanded d-none d-md-block"><!-- d-* hiddens the Sidebar in smaller devices. Its itens can be kept on the Navbar 'Menu' -->
-      <!-- Bootstrap List Group -->
-      <ul class="list-group">
-          <a href="#" data-toggle="sidebar-collapse" class="bg-dark list-group-item list-group-item-action d-flex align-items-center">
-              <div class="d-flex w-100 justify-content-start align-items-center">
-                  <span id="collapse-icon" class="fa fa-2x mr-3"></span>
-                  <span id="collapse-text" class="menu-collapsed">Collapse</span>
-              </div>
-          </a>
-          <!-- settings menu -->
-          <a href="{{ url_for('overview.institute_settings', institute_id=institute._id) }}" class="bg-dark list-group-item list-group-item-action flex-column align-items-start" onclick="hide_div(1)">
-              <div class="d-flex w-100 justify-content-start align-items-center">
-                  <span class="fa fa-wrench fa-fw mr-3"></span>
-                  <span class="menu-collapsed">General settings</span>
-              </div>
-          </a>
-
-          <!-- users menu -->
-          <a href="{{ url_for('overview.institute_users', institute_id=institute._id) }}" class="bg-dark list-group-item list-group-item-action flex-column align-items-start" onclick="hide_div(2)">
-              <div class="d-flex w-100 justify-content-start align-items-center">
-                  <span class="fa fa-users fa-fw mr-3"></span>
-                  <span class="menu-collapsed">Users</span>
-              </div>
-          </a>
-
-          <!-- shared data -->
-          <a href="#submenu1" data-toggle="collapse" aria-expanded="false" class="bg-dark list-group-item list-group-item-action flex-column align-items-start" onclick="hide_div(3)">
-              <div class="d-flex w-100 justify-content-start align-items-center">
-                  <span class="fa fa-share-alt mr-3"></span>
-                  <span class="menu-collapsed">Shared data</span>
-              </div>
-          </a>
-      </ul><!-- List Group END-->
-  </div><!-- sidebar-container END -->
-{% endmacro %}
-
-
 {% macro institute_settings(form, institute) %}
 <form action="{{ url_for('overview.institute_settings', institute_id=institute._id) }}" method="POST"
   accept-charset="utf-8" id="institute_form">
@@ -141,7 +102,7 @@
 
 
 {% macro institute_macro(institute, current_user) %}
-<form action="{{ url_for('overview.institute', institute_id=institute.internal_id) }}">
+<form action="{{ url_for('overview.institute_settings', institute_id=institute.internal_id) }}">
 <div class="card mt-3">
   <div>
   <h5 class="card-title"><a href="{{ url_for('cases.cases', institute_id=institute.internal_id) }}"> {{ institute.display_name }}</a>
